View manager registration and store registration and  store address were combined
I found some bugs
In the previous version, the shop registration process was in a chain mode, that is, first the manager registered, then he went to the page of adding shop information, and then he went to the page of adding the address.
If the user intended to be mischievous and stopped in the middle of the registration process, a new record would be added in the database, and this is not good.