APIView and LocalStorage were added to the project to create an order detail in App Orders
In this case
I implemented a series of changes and created a two-state mode. If you are logged in, we have no problem with you and you can place your order.
This is the creation of the order by local storage and sending the required information to the API, which performs this process of creating the orders.
But if you are not logged in, the product information and its quantity will go to View API, where the desired product will be displayed to the user with the help of Ajax.
If you are not logged in, you must log in and after logging in, you will go to the Ajax page that shows the information to the user, and you can create a new order by pressing the submit button.